[feature]: delete user

+++ src/main/java/com/julionepomcueno/workshopspringboot3jpa/services/UserService.java
@@ -2,8 +2,10 @@ package com.julionepomcueno.workshopspringboot3jpa.services;
 
 import com.julionepomcueno.workshopspringboot3jpa.entities.User;
 import com.julionepomcueno.workshopspringboot3jpa.repositories.UserRepository;
+import com.julionepomcueno.workshopspringboot3jpa.services.exceptions.DatabaseException;
 import com.julionepomcueno.workshopspringboot3jpa.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +30,18 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        }
+        catch (DataIntegrityViolationException  e) {
+            throw new DatabaseException(e.getMessage());
+        }
+
+
     }
 
     public User update(Long id, User obj) {
